+++ src/main/kotlin/com/wafflestudio/csereal/core/research/service/ResearchEventService.kt
@@ -1,0 +1,13 @@
+package com.wafflestudio.csereal.core.research.service
+
+import org.springframework.stereotype.Service
+
+interface ResearchEventService {
+}
+
+@Service
+class ResearchEventServiceImpl (
+
+): ResearchEventService {
+
+}
