+++ src/main/kotlin/com/wafflestudio/csereal/core/research/service/ResearchSearchService.kt
@@ -18,10 +18,6 @@ import org.springframework.transaction.annotation.Propagation
 import org.springframework.transaction.annotation.Transactional
 
 interface ResearchSearchService {
-    fun professorCreatedEventListener(professorCreatedEvent: ProfessorCreatedEvent)
-    fun professorDeletedEventListener(professorDeletedEvent: ProfessorDeletedEvent)
-    fun professorModifiedEventListener(professorModifiedEvent: ProfessorModifiedEvent)
-    fun deleteResearchSearch(researchSearchEntity: ResearchSearchEntity)
     fun searchTopResearch(keyword: String, language: LanguageType, number: Int, amount: Int): ResearchSearchResBody
     fun searchResearch(
         keyword: String,
@@ -30,6 +26,7 @@ interface ResearchSearchService {
         pageNum: Int,
         amount: Int
     ): ResearchSearchResBody
+    fun deleteResearchSearch(researchSearchEntity: ResearchSearchEntity)
 }
 
 @Service
@@ -71,58 +68,6 @@ class ResearchSearchServiceImpl(
                 total = it.second
             )
         }
-
-    @EventListener
-    @Transactional
-    override fun professorCreatedEventListener(professorCreatedEvent: ProfessorCreatedEvent) {
-        val lab = professorCreatedEvent.labId?.let {
-            labRepository.findByIdOrNull(it)
-        } ?: return
-
-        lab.researchSearch?.update(lab) ?: let {
-            lab.researchSearch = ResearchSearchEntity.create(lab)
-        }
-    }
-
-    @EventListener
-    @Transactional
-    override fun professorDeletedEventListener(professorDeletedEvent: ProfessorDeletedEvent) {
-        val lab = professorDeletedEvent.labId?.let {
-            labRepository.findByIdOrNull(it)
-        } ?: return
-
-        // if lab still has professor, remove it
-        lab.professors.removeIf { it.id == professorDeletedEvent.id }
-
-        // update search data
-        lab.researchSearch?.update(lab)
-    }
-
-    @EventListener
-    @Transactional
-    override fun professorModifiedEventListener(professorModifiedEvent: ProfessorModifiedEvent) {
-        val beforeLab = professorModifiedEvent.beforeLabId?.let {
-            labRepository.findByIdOrNull(it)
-        }
-
-        val afterLab = professorModifiedEvent.afterLabId?.let {
-            labRepository.findByIdOrNull(it)
-        }
-
-        if (beforeLab != null && beforeLab == afterLab) {
-            beforeLab.researchSearch?.update(beforeLab)
-        }
-
-        beforeLab?.run {
-            // if lab still has professor, remove it
-            professors.removeIf { it.id == professorModifiedEvent.id }
-            researchSearch?.update(this)
-        }
-
-        afterLab?.run {
-            researchSearch?.update(this)
-        }
-    }
 
     @EventListener
     @Transactional(propagation = Propagation.REQUIRES_NEW)
